requirements.txt
altair==4.1.0; python_version >= '3.6'
argon2-cffi==20.1.0
astor==0.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
async-generator==1.10; python_version >= '3.5'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
base58==2.1.0; python_version >= '3.5'
bleach==3.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
blinker==1.4
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorama==0.4.4; sys_platform == 'win32'
cycler==0.10.0
decorator==5.0.9; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints==0.3; python_version >= '2.7'
et-xmlfile==1.1.0; python_version >= '3.6'
gitdb==4.0.7; python_version >= '3.4'
gitpython==3.1.17; python_version >= '3.5'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ipykernel==5.5.5; python_version >= '3.4'
ipython-genutils==0.2.0
ipython==7.24.1; python_version >= '3.3'
ipywidgets==7.6.3
jedi==0.18.0; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
jsonschema==3.2.0
jupyter-client==6.1.12; python_version >= '3.5'
jupyter-core==4.7.1; python_version >= '3.6'
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0; python_version >= '3.6'
kiwisolver==1.3.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
matplotlib-inline==0.1.2; python_version >= '3.5'
matplotlib==3.4.2; python_version >= '3.7'
mistune==0.8.4
nbclient==0.5.3; python_full_version >= '3.6.1'
nbconvert==6.0.7; python_version >= '3.6'
nbformat==5.1.3; python_version >= '3.5'
nest-asyncio==1.5.1; python_version >= '3.5'
notebook==6.4.0; python_version >= '3.6'
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2; python_version >= '3.6'
pickleshare==0.7.5
pillow==8.2.0
plotly==4.14.3
prometheus-client==0.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prompt-toolkit==3.0.18; python_full_version >= '3.6.1'
protobuf==3.17.2
pyarrow==4.0.1; python_version < '3.9'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydeck==0.6.2
pygments==2.9.0; python_version >= '3.5'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrsistent==0.17.3; python_version >= '3.5'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pywin32==301; sys_platform == 'win32'
pywinpty==1.1.1; os_name == 'nt'
pyzmq==22.1.0; python_version >= '3.6'
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
retrying==1.3.3
scipy==1.6.3; python_version < '3.10' and python_version >= '3.7'
seaborn==0.11.1
send2trash==1.5.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==4.0.0; python_version >= '3.5'
streamlit==0.82.0
terminado==0.10.0; python_version >= '3.6'
testpath==0.5.0; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.11.1; python_version >= '3.5'
tornado==6.1; python_version >= '3.5'
traitlets==5.0.5; python_version >= '3.7'
tzlocal==2.1
urllib3==1.26.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
validators==0.18.2; python_version >= '3.4'
watchdog==2.1.2; platform_system != 'Darwin'
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1